+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # Changelog for pinky
 
+## Unreleased changes
+
+### Added
+
+- `fromBf :: Bf -> BrainfuckM ()`
+
+### Changed
+
+- Rename `trimTape` -> `truncateTape` and move into `BrainfuckTape` module
+
 ## 0.1.0
 
 - Initial release